creating apporpiate files for testing
